backend/requirement.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
python-decouple>=3.8,<4.0
pytz>=2023.3,<2024.0
sqlparse>=0.4.4,<1.0
psycopg2-binary>=2.9,<3.0 
django-simple-history>=3.0,<4.0
django-allauth>=0.59,<1.0
django-rest-auth>=0.9.5,<1.0
gunicorn>=21.2,<22.0
Pillow
django-ckeditor-5